requirements.txt
addict==2.4.0
ansi2html==1.8.0
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.2.0
cycler==0.12.1
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
fastjsonschema==2.19.0
Flask==3.0.0
fonttools==4.44.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
jupyter_core==5.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
nbformat==5.7.0
nest-asyncio==1.5.8
numpy==1.26.2
open3d==0.17.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
retrying==1.3.4
rpds-py==0.12.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
